+++ lib/clixon/clixon_nacm.h
@@ -57,7 +57,7 @@ enum nacm_access{
  * Prototypes
  */
 int nacm_proxyuser_add(clixon_handle h, const char *user);
-int nacm_rpc(const char *rpc, const char *module, const char *username, cxobj *xnacm, cbuf *cbret);
+int nacm_rpc(clixon_handle h, const char *rpc, const char *module, const char *username, cxobj *xnacm, cbuf *cbret);
 int nacm_datanode_read1(clixon_handle h, cxobj *xt, const char *username, cxobj *nacm_xtree);
 int nacm_datanode_read_prune(clixon_handle h, cxobj *xt);
 int nacm_datanode_write(clixon_handle h, cxobj *xr, cxobj *xt,